feat: Implement student attempts provider and associated APIs
- Added `MyAttemptDto` and related DTOs for managing quiz attempts.
- Created `StudentAttemptsProvider` to manage student attempts state and provide context.
- Introduced `assignedQuizAvailability.ts` to handle quiz availability logic.
- Developed quiz adapters in `quizzesAdapters.ts` for mapping quiz data.
- Implemented quiz generation API in `quizGenerationApi.ts` for creating quizzes from text and PDF.
- Added attempt-related APIs in `attemptsApi.ts` for starting and submitting attempts.
- Created utility functions in `attemptsUtils.ts` for checking attempt eligibility and progress.
- Developed hooks `useQuizAttempts` and `useQuizAttemptsMap` for managing quiz attempts state in components.
- Updated `TeacherClassesPage` and `TeacherQuizLibraryPage` to handle quiz removal and deletion with error handling.
- Added `.gitignore` to exclude unnecessary files from version control.

+++ backend/Bilgenly.Application/DTOs/ClassAnalyticsDto.cs
@@ -14,7 +14,7 @@ public class ClassAnalyticsDto
     public int NeedsAttentionCount { get; set; }   
     public double CompletionRate { get; set; }
     public double? AverageScore { get; set; }
-    public int AvgAttemptsUsed { get; set; }
+    public double AvgAttemptsUsed { get; set; }
     public DateTime? Deadline { get; set; }
     public int? MaxAttempts { get; set; }
     public List<StudentAssignmentResultDto> StudentResults { get; set; } = new();
